Breaks out league statistics methodsfrom stat_tracker
