+++ lib/league_statistics.rb
@@ -1,0 +1,89 @@
+module LeagueStatistics
+
+  def count_of_teams
+    @teams.all.count
+  end
+
+  def best_offense
+    highest_scoring = goals_scored_by_team.max_by do |team_id, total_goals|
+      total_goals
+    end
+    @teams.find_by_id(highest_scoring.first).team_name
+  end
+
+  def worst_offense
+    lowest_scoring = goals_scored_by_team.min_by do |team_id, total_goals|
+      total_goals
+    end
+    @teams.find_by_id(lowest_scoring.first).team_name
+  end
+
+  def best_defense
+    least_allowed = goals_allowed_by_team.min_by do |team_id, total_goals|
+      total_goals
+    end
+    @teams.find_by_id(least_allowed.first).team_name
+  end
+
+  def worst_defense
+    most_allowed = goals_allowed_by_team.max_by do |team_id, total_goals|
+      total_goals
+    end
+    @teams.find_by_id(most_allowed.first).team_name
+  end
+
+  def highest_scoring_visitor
+    teams_away_goals = goals_for_visitors
+     highest_team = teams_away_goals.max_by {|team, average_goals| average_goals}
+     @teams.find_by_id(highest_team.first).team_name
+  end
+
+  def highest_scoring_home_team
+    teams_home_goals = goals_for_home_teams
+    highest_team = teams_home_goals.max_by {|team, average_goals| average_goals}
+    @teams.find_by_id(highest_team.first).team_name
+  end
+
+  def lowest_scoring_visitor
+    teams_away_goals = goals_for_visitors
+    highest_team = teams_away_goals.min_by {|team, average_goals| average_goals}
+    @teams.find_by_id(highest_team.first).team_name
+  end
+
+  def lowest_scoring_home_team
+    teams_home_goals = goals_for_home_teams
+    highest_team = teams_home_goals.min_by {|team, average_goals| average_goals}
+    @teams.find_by_id(highest_team.first).team_name
+  end
+
+  def winningest_team
+    team_games = group_games_by_team
+    team_games.each do |team_id, games|
+      team_games[team_id] = calculate_win_percentage(team_id, games)
+    end
+    @teams.find_by_id((team_games.max_by {|team, goals| goals}).first).team_name
+  end
+
+  def best_fans
+    teams_wins = {}
+    @teams.all.each do |team|
+      teams_wins[team] = calc_home_win_percentages(team.id, @games.all) - calc_away_win_percentages(team.id, @games.all)
+    end
+    best_fans = teams_wins.max_by do |team, percentages|
+      percentages
+    end
+    best_fans.first.team_name
+  end
+
+  def worst_fans
+    teams_wins = {}
+    @teams.all.each do |team|
+      teams_wins[team] = calc_home_win_percentages(team.id, @games.all) - calc_away_win_percentages(team.id, @games.all)
+    end
+    worst_fans = teams_wins.find_all do |team, percentages|
+      percentages < 50
+    end
+    worst_fans.map {|team| team.first.team_name}
+  end
+
+end
